refactor: move config from querylimit package to repository and update related imports
