+++ repository/req_test.go
@@ -11,7 +11,6 @@ import (
 
 	grpcclient "github.com/dezh-tech/immortal/infrastructure/grpc_client/gen"
 	infra "github.com/dezh-tech/immortal/infrastructure/meilisearch"
-	"github.com/dezh-tech/immortal/repository/querylimit"
 	"github.com/dezh-tech/immortal/types"
 	"github.com/dezh-tech/immortal/types/event"
 	"github.com/dezh-tech/immortal/types/filter"
@@ -257,7 +256,7 @@ func TestHandleReq(t *testing.T) {
 		db:     nil,
 		meili:  meili,
 		grpc:   mockGRPC,
-		config: &querylimit.Config{},
+		config: &Config{},
 	}
 	handler.config.SetDefaultQueryLimit(0)
 	handler.config.SetMaxQueryLimit(10)
